Add debian 9 to 10 script

+++ voxscript.sh
@@ -13,6 +13,7 @@ if "$lang" == "en":
 	echo "1- Update Debian 8 to 10 (deb8to10.sh)"
 	echo "2- Install flood seedbox in Debian 10 (seedbox_deb10_flood.sh)"
 	echo "3- Update CentOS 7 to 8 (centos7to8.sh)"
+	echo "4- Update Debian 9 to 10 (deb9to10.sh)"
 	read -p "What's your choices ?" chx
 	if "$chx" == "1":
 		mkdir /tmp/vscript
@@ -35,6 +36,13 @@ if "$lang" == "en":
 		bash /tmp/vscript/centos7to8.sh
 		rm -r /tmp/vscript
 		exit
+	if "$chx" == "4":
+		mkdir /tmp/vscript
+		curl https://voxan24.github.io/Linux_Script/deb9to10.sh -o /tmp/vscript/deb9t10.sh
+		chmod a+x /tmp/vscript/deb9t10.sh
+		bash /tmp/vscript/deb9t10
+		rm -r /tmp/vscript
+		exit
 	else:
 		echo "Invalid, goodbye"
 		exit
@@ -45,6 +53,7 @@ if "lang" == "fr":
 	echo "1- Script de mise a niveaux de Debian 8 vers Debian 10 (deb8to10.sh)"
 	echo "2- Script d'installation d'une seedbox avec flood sous Debian 10 (seedbox_deb10_flood.sh)"
 	echo "3- Script de mise a niveaux de CentOS 7 vers CentOS 8 (centos7to8.sh)"
+	echo "4- Script de mise a niveaux de Debian 9 vers Debian 10 (deb9to10.sh)"
 	read -p "Quel est votre choix (entrez le chiffre correspondant a votre choix)" chx
 	if "$chx" == "1":
 		mkdir /tmp/vscript
@@ -65,6 +74,13 @@ if "lang" == "fr":
 		curl https://voxan24.github.io/Linux_Script/centos7to8.sh -o /tmp/vscript/centos7to8.sh
 		chmod a+x /tmp/vscript/centos7to8.sh
 		bash /tmp/vscript/centos7to8.sh
+		rm -r /tmp/vscript
+		exit
+	if "$chx" == "4":
+		mkdir /tmp/vscript
+		curl https://voxan24.github.io/Linux_Script/deb9to10.sh -o /tmp/vscript/deb9t10.sh
+		chmod a+x /tmp/vscript/deb9t10.sh
+		bash /tmp/vscript/deb9t10
 		rm -r /tmp/vscript
 		exit
 	else:
